refactor crystallize shield and shard
