+++ pkg/conditional/gadgets.go
@@ -3,11 +3,11 @@ package conditional
 import (
 	"fmt"
 
+	"github.com/genshinsim/gcsim/internal/template/crystallize"
 	"github.com/genshinsim/gcsim/internal/template/dendrocore"
 	"github.com/genshinsim/gcsim/internal/template/sourcewaterdroplet"
 	"github.com/genshinsim/gcsim/pkg/core"
 	"github.com/genshinsim/gcsim/pkg/core/attributes"
-	"github.com/genshinsim/gcsim/pkg/reactable"
 )
 
 func evalGadgets(c *core.Core, fields []string) (int, error) {
@@ -61,7 +61,7 @@ func evalCrystallizeShard(c *core.Core, key string) (int, error) {
 	case "all":
 		count := 0
 		for _, g := range c.Combat.Gadgets() {
-			cs, ok := g.(*reactable.CrystallizeShard)
+			cs, ok := g.(*crystallize.Shard)
 			if !ok {
 				continue
 			}
@@ -81,7 +81,7 @@ func evalCrystallizeShard(c *core.Core, key string) (int, error) {
 func countElementalCrystallizeShards(c *core.Core, ele string) int {
 	count := 0
 	for _, g := range c.Combat.Gadgets() {
-		cs, ok := g.(*reactable.CrystallizeShard)
+		cs, ok := g.(*crystallize.Shard)
 		if !ok {
 			continue
 		}
